# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Score API",
        "contact": {
            "name": "Score API Developer Team",
            "url": "https://github.com/pine09/ScoreAPI"
        },
        "version": "1.0"
    },
    "host": "score.app",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/v1/admin/dosen": {
            "get": {
                "tags": [
                    "Admin on Dosen"
                ],
                "summary": "Mengambil semua data dosen yang ada.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data dosen. (max 5 per page)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dosen"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin on Dosen"
                ],
                "summary": "Mendaftarkan dosen baru.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Data Dosen Baru.",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dosen"
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data dosen baru."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    },
                    "409": {
                        "description": "Conflict."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    }
                }
            }
        },
        "/api/v1/admin/dosen/{id}": {
            "get": {
                "tags": [
                    "Admin on Dosen"
                ],
                "summary": "Mengambil data dosen spesifik",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data dosen yang dicari"
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    },
                    "404": {
                        "description": "Resource not found."
                    }
                }
            },
            "put": {
                "tags": [
                    "Admin on Dosen"
                ],
                "summary": "Mengedit data diri dosen.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Data Dosen (Edit).",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dosen"
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data dosen edited."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    },
                    "404": {
                        "description": "Resource not found."
                    },
                    "409": {
                        "description": "Conflict."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "500": {
                        "description": "Internal Server Error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin on Dosen"
                ],
                "summary": "Men-delete data dosen.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data dosen deleted."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "403": {
                        "description": "Forbidden action."
                    },
                    "404": {
                        "description": "Resource not found."
                    },
                    "422": {
                        "description": "Unprocessable Entity."
                    },
                    "500": {
                        "description": "Internal Server Error."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dosen": {
            "properties": {
                "NIDN": {
                    "type": "string"
                },
                "foto": {
                    "description": "path to image. nullable field",
                    "type": "string"
                },
                "nama_depan": {
                    "type": "string"
                },
                "nama_belakang": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "alamat": {
                    "type": "text"
                },
                "jenis_kelamin": {
                    "type": "enum('Laki-Laki', 'Perempuan')"
                }
            }
        },
...
    }
}